move logger functions out of runner.cpp
